feat: restructure components and implement new About, Contact, Header, PersonalIntroduction, Projects, and TechnicalSkillsAndDegree components
