+++ src/App.tsx
@@ -1,17 +1,18 @@
 
 import React from 'react';
-import Header from './components/Header';
-import About from './components/About';
-import PersonalIntroduction from './components/PersonalIntroduction';
-import Projects from './components/Projects';
-import TechnicalSkillsAndDegree from './components/TechnicalSkillsAndDegree';
-import Contact from './components/Contact';
+import Header from './components/header/Header';
+
+import About from './components/about/About';
+import PersonalIntroduction from './components/personalIntroduction/PersonalIntroduction';
+import Projects from './components/projects/Projects';
+import TechnicalSkillsAndDegree from './components/technicalSkillsAndDegree/TechnicalSkillsAndDegree';
+import Contact from './components/contact/Contact';
 
 const App: React.FC = () => {
   return (
-    <div className="min-h-screen w-full bg-white">
+    <div className="bg-white">
       <Header />
-      <main className="min-h-screen w-full mx-auto">
+      <main className="">
         <PersonalIntroduction />
         <About />
         <Projects />
