+++ src/components/header/Header.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 
 const Header: React.FC = () => {
   return (
-    <header className="bg-slate-100 p-4">
-      <div className="container mx-auto flex justify-between items-center">
-        <h1 className="text-2xl font-bold text-blue-800">Lucas Lotar</h1>
+    <header className="w-full px-4 py-2 bg-slate-100">
+      <div className="flex justify-between items-center">
+        <h1 className="text-2xl font-bold text-blue-800 mx-26">Lucas Lotar</h1>
 
-        <nav>
-          <ul className="flex space-x-4">
+        <nav className="mx-26">
+          <ul className="flex space-x-20">
             <li><a href="#sobre" className="hover:text-gray-300">Sobre</a></li>
 
             <li><a href="#projetos" className="hover:text-gray-300">Projetos</a></li>
