+++ src/components/header/header.css
@@ -1,0 +1,6 @@
+/* .header {
+  background-color: red;
+  color: white;
+  padding: 10px;
+  border: 2px solid black;
+} */
