+++ src/components/technicalSkillsAndDegree/TechnicalSkillsAndDegree.tsx
